작업1

+++ Advanced Calc/Divi.cpp
@@ -1,6 +1,6 @@
 
 #include "Command.h"
 
-int Divi(int a, int b) {
+double Divi(int a, int b) {
 	return a / b;
 }
