+++ Advanced Calc/Minus.cpp
@@ -1,6 +1,6 @@
 
 #include "Command.h"
 
-int Minus(int a, int b) {
+double Minus(int a, int b) {
 	return a - b;
 }
