+++ Advanced Calc/Mult.cpp
@@ -3,6 +3,6 @@
 #include "Command.h"
 
 
-int Mult(int a, int b) { 
+double Mult(int a, int b) { 
 	return a * b; 
 }
